Wallet API remove app config
