+++ wallet/src/main/scala/org/bitcoins/wallet/api/HDWalletApi.scala
@@ -436,8 +436,7 @@ trait HDWalletApi extends WalletApi {
     } yield tx
   }
 
-  def listDefaultAccountUtxos(): Future[Vector[SpendingInfoDb]] =
-    listUtxos(walletConfig.defaultAccount)
+  def listDefaultAccountUtxos(): Future[Vector[SpendingInfoDb]]
 
   def listUtxos(account: HDAccount): Future[Vector[SpendingInfoDb]]
